+++ include/los/event.h
@@ -1,17 +1,17 @@
 #ifndef __LOS_EVENT_H
 #define __LOS_EVENT_H
 
-#include <stdint.h>
+#include "types.h"
 
 #define EVENT_TYPE_KEY_PRESS 0
 #define EVENT_TYPE_KEY_RELEASE 1
 
 typedef struct {
-    uint64_t type;
-    uint64_t param1;
-    uint64_t param2;
+    usize type;
+    usize param1;
+    usize param2;
 } Event;
 
-int64_t peek_event(Event* pevent);
+isize peek_event(Event* pevent);
 
 #endif
