+++ include/los/time.h
@@ -1,10 +1,10 @@
 #ifndef __LOS_TIME_H
 #define __LOS_TIME_H
 
-#include <stdint.h>
+#include "types.h"
 
-int64_t get_process_time();
-int64_t get_time_zone();
-int64_t get_epoch_time();
+isize get_process_time();
+isize get_time_zone();
+isize get_epoch_time();
 
 #endif
