+++ src/console/console_clear.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_clear() { return system_call0(CONSOLE_CLEAR_SYSCALL); }
+isize console_clear() { return system_call0(CONSOLE_CLEAR_SYSCALL); }
