+++ src/console/console_get_height.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_get_height() { return system_call0(CONSOLE_GET_HEIGHT_SYSCALL); }
+isize console_get_height() { return system_call0(CONSOLE_GET_HEIGHT_SYSCALL); }
