+++ src/console/console_set_attribute.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_set_attribute(uint64_t attribute) { return system_call1(CONSOLE_SET_ATTRIBUTE_SYSCALL, attribute); }
+isize console_set_attribute(usize attribute) { return system_call1(CONSOLE_SET_ATTRIBUTE_SYSCALL, attribute); }
