+++ src/console/console_set_foreground.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_set_foreground(uint64_t color) { return system_call1(CONSOLE_SET_FOREGROUND_SYSCALL, color); }
+isize console_set_foreground(usize color) { return system_call1(CONSOLE_SET_FOREGROUND_SYSCALL, color); }
