+++ src/console/console_write_str.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_write_str(const char* string) { return system_call1(CONSOLE_WRITE_STR_SYSCALL, (uint64_t)string); }
+isize console_write_str(const char* string) { return system_call1(CONSOLE_WRITE_STR_SYSCALL, (usize)string); }
