+++ src/filesystem/remove_file.c
@@ -2,6 +2,6 @@
 
 #include <los/syscall.h>
 
-int64_t remove_file(const char* path) {
-    return system_call1(REMOVE_FILE_SYSCALL, (uint64_t)path);
+isize remove_file(const char* path) {
+    return system_call1(REMOVE_FILE_SYSCALL, (usize)path);
 }
