+++ src/process/exit_process.c
@@ -2,8 +2,8 @@
 
 #include <los/syscall.h>
 
-void exit_process(int64_t status) {
-    system_call1(EXIT_PROCESS_SYSCALL, (uint64_t)status);
+void exit_process(isize status) {
+    system_call1(EXIT_PROCESS_SYSCALL, (usize)status);
     while (1)
         ;
 }
