+++ src/thread/create_thread.c
@@ -16,9 +16,9 @@ extern void thread_entry(void* context);
 
 ThreadID create_thread(ThreadEntry entry) {
     ThreadEntryContext* context = (ThreadEntryContext*)allocate_memory(sizeof(ThreadEntryContext), 8);
-    context->stack_top = (void*)(((uint64_t)allocate_memory(STACK_SIZE, 16)) + STACK_SIZE);
+    context->stack_top = (void*)(((usize)allocate_memory(STACK_SIZE, 16)) + STACK_SIZE);
     context->tls = create_tls();
     context->entry = entry;
 
-    return system_call2(CREATE_THREAD_SYSCALL, (uint64_t)thread_entry, (uint64_t)context);
+    return system_call2(CREATE_THREAD_SYSCALL, (usize)thread_entry, (usize)context);
 }
