+++ src/thread/exit_thread.c
@@ -5,12 +5,12 @@
 
 extern void* get_fs_base();
 
-void exit_thread(int64_t status) {
+void exit_thread(isize status) {
     ThreadStruct* ptr = get_fs_base();
 
     destroy_tls(ptr);
 
-    system_call1(EXIT_THREAD_SYSCALL, (uint64_t)status);
+    system_call1(EXIT_THREAD_SYSCALL, (usize)status);
 
     while (1)
         ;
